requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
-e file:../odp-core
    # via -r requirements.in
alembic==1.13.1
    # via -r requirements.in
anyio==4.3.0
    # via
    #   httpx
    #   starlette
async-timeout==4.0.3
    # via redis
authlib==1.3.0
    # via odp
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
coverage==7.4.3
    # via -r requirements.in
cryptography==42.0.5
    # via authlib
exceptiongroup==1.2.0
    # via
    #   anyio
    #   pytest
factory-boy==3.3.0
    # via -r requirements.in
faker==24.1.0
    # via
    #   -r requirements.in
    #   factory-boy
fastapi==0.110.0
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via -r requirements.in
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
mako==1.3.2
    # via alembic
markupsafe==2.1.5
    # via mako
ory-hydra-client==1.11.8
    # via odp
packaging==24.0
    # via pytest
pluggy==1.4.0
    # via pytest
psycopg2==2.9.9
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic[dotenv]==1.10.14
    # via
    #   fastapi
    #   odp
pytest==8.1.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   faker
    #   ory-hydra-client
python-dotenv==1.0.1
    # via pydantic
redis==5.0.3
    # via odp
requests==2.31.0
    # via odp
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.28
    # via
    #   -r requirements.in
    #   alembic
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.1
    # via -r requirements.in
starlette==0.36.3
    # via
    #   -r requirements.in
    #   fastapi
tomli==2.0.1
    # via pytest
typing-extensions==4.10.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   sqlalchemy
    #   uvicorn
urllib3==2.2.1
    # via
    #   ory-hydra-client
    #   requests
uvicorn==0.28.0
    # via -r requirements.in
